CHANGED THE LOGIC OF THE CODE
Thinking about implementing other path finding algorithms, it may be
better to modify the expand_node function so that it can work with all
possible algorithms. So I left the duty to compute the approximate cost
and so the heuristic, to the algorithms that need it. So I compute the
approximate cost inside the A_star function.