ERROR IN CHECKING THE MEMORY
I noticed that the number of objects in memory, when the solution is
found, was senseless. That was becouse I checked it after exiting the
function and the garbage collector, of course, deleated all unnecessary objects that were
not part of the solution. Only the objects in the target state path remained since
that was the only object returned by the function.
That's why I put the memory usage check when the solution is found and
before exiting the function.